fix: 알림 관련 type 분리 및 알림 토스트 클릭 시 단일알림 조회하도록 수정
