+++ front/app/(route)/schedule/components/ScheduleAddForm.tsx
@@ -60,7 +60,7 @@ const ScheduleAddForm = ({
     scheduleType: 'WALK',
     mates: [],
     scheduleDate: formatDateToYMD(selectedDateFromCalender),
-    scheduleTime: '',
+    scheduleTime: '20:35',
     repeatType: 'NONE',
     alertType: 'NONE',
     memo: '',
@@ -78,7 +78,7 @@ const ScheduleAddForm = ({
         scheduleType: loadedScheduleData.scheduleType || 'WALK',
         mates: formattedMates,
         scheduleDate: parseDateToYMD(loadedScheduleData.scheduleDate),
-        scheduleTime: formatTimeToHM(loadedScheduleData.scheduleTime) || '',
+        scheduleTime: formatTimeToHM(loadedScheduleData.scheduleTime) || '20:35',
         repeatType: loadedScheduleData.repeatType || 'NONE',
         alertType: loadedScheduleData.alertType || 'NONE',
         memo: loadedScheduleData.memo || '',
@@ -167,7 +167,7 @@ const ScheduleAddForm = ({
       scheduleType: 'WALK',
       mates: [],
       scheduleDate: formatDateToYMD(selectedDateFromCalender),
-      scheduleTime: '',
+      scheduleTime: '20:35',
       repeatType: 'NONE',
       alertType: 'NONE',
       memo: '',
